revision 4

+++ app/Http/Controllers/CommentController.php
@@ -4,14 +4,21 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function create()
     {
         $comment = new Comment;
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
         return back();
     }
@@ -19,7 +26,11 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
-        return back();
+        if(Gate::allows('comment-delete',$comment))
+        {
+            $comment->delete();
+            return back();
+        }
+        return back()->with('error','Unauthorize');
     }
 }
